Moving test to solvers
