+++ .github/workflows/ci_cd.yml
@@ -91,61 +91,56 @@ jobs:
 # # vvvvvvvvvvvvvvvvvvvvvvvvvvvvvv    RUNNING ON SELF-HOSTED RUNNER    vvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
 # # =================================================================================================
 
-#  system-tests-visualization-windows:
-#    name: Testing and coverage Visualization (Windows)
-#    needs: [integration-tests]
-#    if: github.event.pull_request.draft == false
-#    runs-on: [ self-hosted, Windows, pyaedt ]
-#    steps:
-#      - name: Install Git and checkout project
-#        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-#
-#      - name: Setup Python
-#        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5.6.0
-#        with:
-#          python-version: ${{ env.MAIN_PYTHON_VERSION }}
-#
-#      - name: Create virtual environment
-#        run: |
-#          python -m venv .venv
-#          .venv\Scripts\Activate.ps1
-#          python -m pip install pip -U
-#          python -m pip install wheel setuptools -U
-#          python -c "import sys; print(sys.executable)"
-#
-#      - name: Install pyaedt and tests dependencies
-#        run: |
-#          .venv\Scripts\Activate.ps1
-#          pip install .[tests]
-#
-#      - name: Install CI dependencies (e.g. vtk-osmesa)
-#        run: |
-#          .venv\Scripts\Activate.ps1
-#          # Uninstall conflicting dependencies
-#          pip uninstall --yes vtk
-#          pip install --extra-index-url https://wheels.vtk.org vtk-osmesa
-#
-#      - name: Run tests marked with 'visualization'
-#        uses: nick-fields/retry@ce71cc2ab81d554ebbe88c79ab5975992d79ba08 # v3.0.2
-#        env:
-#          PYTHONMALLOC: malloc
-#        with:
-#          max_attempts: 2
-#          retry_on: error
-#          timeout_minutes: 120
-#          command: |
-#            .venv\Scripts\Activate.ps1
-#            pytest ${{ env.PYTEST_ARGUMENTS }} -n 4 --dist loadfile --timeout=600 -m visualization
-#
-#      - uses: codecov/codecov-action@ad3126e916f78f00edff4ed0317cf185271ccc2d # v5.4.2
-#        with:
-#          name: codecov-system-visualization-tests-windows
-#          files: ./coverage.xml
-#          flags: windows_system_visualization
-#
-#      - name: Upload pytest test results
-#        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
-#        with:
-#          name: pytest-visualization-windows
-#          path: junit/test-results.xml
-#        if: ${{ always() }}
+  system-tests-solvers-windows:
+    name: Testing solvers and coverage (Windows)
+    needs: [integration-tests]
+    if: github.event.pull_request.draft == false
+    runs-on: [ self-hosted, Windows, pyaedt ]
+    steps:
+      - name: Install Git and checkout project
+        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+
+      - name: Setup Python
+        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5.6.0
+        with:
+          python-version: ${{ env.MAIN_PYTHON_VERSION }}
+
+      - name: Create virtual environment
+        run: |
+          python -m venv .venv
+          .venv\Scripts\Activate.ps1
+          python -m pip install pip -U
+          python -m pip install wheel setuptools -U
+          python -c "import sys; print(sys.executable)"
+
+      - name: Install pyaedt and tests dependencies
+        run: |
+          .venv\Scripts\Activate.ps1
+          pip install .[tests]
+
+      - name: Install CI dependencies (e.g. vtk-osmesa)
+        run: |
+          .venv\Scripts\Activate.ps1
+          # Uninstall conflicting dependencies
+          pip uninstall --yes vtk
+          pip install --extra-index-url https://wheels.vtk.org vtk-osmesa
+
+      - name: Run tests marked with 'solvers'
+        env:
+          PYTHONMALLOC: malloc
+        run: |
+          .venv\Scripts\Activate.ps1
+          pytest ${{ env.PYTEST_ARGUMENTS }} --timeout=600 -m solvers --deselect=tests/system/solvers/test_45_FilterSolutions
+
+      - uses: codecov/codecov-action@18283e04ce6e62d37312384ff67231eb8fd56d24 # v5.4.3
+        with:
+          name: codecov-system-solvers-tests-windows
+          files: ./coverage.xml
+          flags: windows_system_solvers
+
+      - name: Upload pytest test results
+        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
+        with:
+          name: pytest-solvers-windows
+          path: junit/test-results.xml
+        if: ${{ always() }}
